+++ reconciler/context.js
@@ -5,23 +5,66 @@
 /*:: import type { Scheduler } from './schedule.js'; */
 
 /*::
-export type ContextState<T> = {|
+export type BranchContext<T> = {|
   value: T,
+  ref: CommitRef,
+|};
+
+export type ContextState<T> = {|
+  subscribers: CommitRef[],
 |};
 
 export type ContextService = {|
-  calculateNextBranch: (element: Element, prev: BranchState) => BranchState,
+  calculateNextBranch: (element: Element, commit: CommitRef, prev: BranchState) => BranchState,
+  calculateNextTargets: (element: Element, prev: Commit) => CommitRef[],
+
+  attachSubscriber: (context: CommitRef, subscriber: CommitRef) => void,
+  detachSubscriber: (context: CommitRef, subscriber: CommitRef) => void,
 |};
 */
 
-export const createContextService = (scheduler/*: Scheduler*/)/*: ContextService*/ => {
-  const calculateNextBranch = (element, prev) => {
+export const createContextService = ()/*: ContextService*/ => {
+  const states = new Map();
+
+  const attachSubscriber = (context, subscriber) => {
+    const prevState = states.get(context.id) || { subscribers: [] };
+
+    if (prevState.subscribers.find(c => c.id === subscriber.id))
+      return;
+
+    const nextState = { subscribers: [...prevState.subscribers, subscriber] };
+
+    states.set(context.id, nextState);
+  }
+  const detachSubscriber = (context, subscriber) => {
+    const prevState = states.get(context.id) || { subscribers: [] };
+
+    const nextState = { subscribers: prevState.subscribers.filter(s => s.id !== subscriber.id) };
+
+    states.set(context.id, nextState);
+  }
+
+  const calculateNextTargets = (element, prev) => {
     const { contextId, value } = (element.props/*: any*/);
 
-    return { ...prev, context: { ...prev.context, [contextId]: { value } } };
+    if (value === prev.element.props.value && contextId === prev.element.props.contextId)
+      return [];
+    
+    const state = states.get(prev.id) || { subscribers: [] };
+    return state.subscribers;
+  }
+
+  const calculateNextBranch = (element, ref, prev) => {
+    const { contextId, value } = (element.props/*: any*/);
+
+    return { ...prev, context: { ...prev.context, [contextId]: { value, ref } } };
   };
 
   return {
+    attachSubscriber,
+    detachSubscriber,
+
     calculateNextBranch,
+    calculateNextTargets,
   };
 };